re-organize types
